+++ python/megane/__init__.py
@@ -85,4 +85,4 @@ __all__ = [
     "view",
     "view_traj",
 ]
-__version__ = "0.12.0"
+__version__ = "0.13.0"